pyproject.toml
[tool.poetry]
name = "movements"
version = "0.1.0"
description = ""
authors = ["Fredrik Gyllenhammar"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.73.0"
uvicorn = {extras = ["standard"], version = "^0.17.4"}
orjson = "^3.6.6"
ujson = "^5.1.0"
email-validator = "^1.1.3"
requests = "^2.27.1"
Jinja2 = "^3.0.3"
python-multipart = "^0.0.5"
itsdangerous = "^2.0.1"
PyYAML = "^6.0"
alembic = "^1.7.6"
ormar = "^0.10.25"

[tool.poetry.dev-dependencies]
flake8-black = "^0.2.4"
mypy = "^0.931"
isort = "^5.10.1"
black = "^22.1.0"
sqlalchemy-stubs = "^0.4"
types-orjson = "^3.6.2"
types-paramiko = "^2.8.16"
types-redis = "^4.1.17"
types-pytz = "^2021.3.5"
types-mock = "^4.0.11"
types-dataclasses = "^0.6.4"
types-contextvars = "^2.4.4"
pydantic = "^1.9.0"
types-python-dateutil = "^2.8.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
skip-string-normalization = true
quiet = true

[tool.isort]
profile = "black"
src_paths = ["movements"]
combine_as_imports = true
line_length = 120
sections = [
    'FUTURE',
    'STDLIB',
    'THIRDPARTY',
    'FIRSTPARTY',
    'LOCALFOLDER'
]
